Call update on touchmove.
In iOS webviews and older versions of iOS Safari, the scroll event will
not fire until you stop scrolling, meaning the header doesn't come back
until the user stops moving, which is unexpected and undesired
behaviour. The touchmove event will make it so that the header comes
back as soon as the user starts scrolling back up, which is what we
want.